delete post

+++ src/main/java/dev/controller/PostCtrl.java
@@ -3,9 +3,12 @@ package dev.controller;
 import dev.controller.dto.PostDTO;
 import dev.entite.forum.Post;
 import dev.service.PostService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @CrossOrigin
 @RestController
@@ -32,8 +35,13 @@ public class PostCtrl {
         return newPostDTO;
     }
 
-    @DeleteMapping
-    public void delete(@RequestBody Integer id) {
-        postService.delete(id);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> delete(@PathVariable Integer id) {
+        Optional<Post> post = postService.findById(id);
+        if(post.isPresent()) {
+            postService.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 }
